created two components for email and sms - pg
