+++ packages/pn-personagiuridica-webapp/src/components/Contacts/EmailContactItem.tsx
@@ -1,10 +1,17 @@
 import { useFormik } from 'formik';
-import { ChangeEvent, useRef, useState } from 'react';
+import { ChangeEvent, useCallback, useEffect, useRef, useState } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import * as yup from 'yup';
 
-import { Box, Button, InputAdornment, Stack, TextField, Typography } from '@mui/material';
-import { CodeModal, DisclaimerModal, ErrorMessage, appStateActions } from '@pagopa-pn/pn-commons';
+import { Box, Button, Stack, TextField, Typography } from '@mui/material';
+import {
+  AppResponse,
+  AppResponsePublisher,
+  CodeModal,
+  DisclaimerModal,
+  ErrorMessage,
+  appStateActions,
+} from '@pagopa-pn/pn-commons';
 import { ButtonNaked } from '@pagopa/mui-italia';
 
 import { AddressType, ChannelType } from '../../models/contacts';
@@ -12,18 +19,12 @@ import { createOrUpdateAddress, deleteAddress } from '../../redux/contact/action
 import { SaveDigitalAddressParams } from '../../redux/contact/types';
 import { useAppDispatch, useAppSelector } from '../../redux/hooks';
 import { RootState } from '../../redux/store';
-import {
-  contactAlreadyExists,
-  emailValidationSchema,
-  internationalPhonePrefix,
-  phoneValidationSchema,
-} from '../../utility/contacts.utility';
+import { contactAlreadyExists, emailValidationSchema } from '../../utility/contacts.utility';
 import DeleteDialog from './DeleteDialog';
 import DigitalContactElem from './DigitalContactElem';
 import ExistingContactDialog from './ExistingContactDialog';
 
 interface Props {
-  type: ChannelType.EMAIL | ChannelType.SMS;
   value: string;
   blockDelete?: boolean;
 }
@@ -35,11 +36,10 @@ enum ModalType {
   DELETE = 'delete',
 }
 
-const CourtesyContactItem = ({ type, value, blockDelete }: Props) => {
+const EmailContactItem = ({ value, blockDelete }: Props) => {
   const { t } = useTranslation(['common', 'recapiti']);
   const digitalAddresses =
     useAppSelector((state: RootState) => state.contactsState.digitalAddresses) ?? [];
-  const contactType = type.toLowerCase();
   const digitalElemRef = useRef<{ editContact: () => void; toggleEdit: () => void }>({
     editContact: () => {},
     toggleEdit: () => {},
@@ -49,30 +49,24 @@ const CourtesyContactItem = ({ type, value, blockDelete }: Props) => {
   const codeModalRef =
     useRef<{ updateError: (error: ErrorMessage, codeNotValid: boolean) => void }>(null);
 
-  const emailValidationSch = yup.object().shape({
+  const validationSchema = yup.object().shape({
     email: emailValidationSchema(t),
   });
 
-  // note that phoneValidationSchema depends on the phoneRegex which is different
-  // for the insertion and modification cases, check the comment
-  // about the useEffect which calls setPhoneRegex below
-  const phoneValidationSch = yup.object().shape({
-    sms: phoneValidationSchema(t, !!value),
-  });
-
   const initialValues = {
-    [contactType]: value ?? '',
+    email: value ?? '',
   };
 
   const formik = useFormik({
     initialValues,
-    validationSchema: type === ChannelType.EMAIL ? emailValidationSch : phoneValidationSch,
+    validationSchema,
     enableReinitialize: true,
     validateOnMount: true,
     onSubmit: () => {
-      const contactValue = getContactValue(type);
       // first check if contact already exists
-      if (contactAlreadyExists(digitalAddresses, contactValue, 'default', type)) {
+      if (
+        contactAlreadyExists(digitalAddresses, formik.values.email, 'default', ChannelType.EMAIL)
+      ) {
         setModalOpen(ModalType.EXISTING);
         return;
       }
@@ -80,21 +74,17 @@ const CourtesyContactItem = ({ type, value, blockDelete }: Props) => {
     },
   });
 
-  const getContactValue = (type: ChannelType) => {
-    if (type === ChannelType.EMAIL) {
-      return formik.values.email;
-    }
-    return type === ChannelType.SMS && value
-      ? formik.values.sms
-      : internationalPhonePrefix + formik.values.sms;
+  const handleChangeTouched = async (e: ChangeEvent) => {
+    formik.handleChange(e);
+    await formik.setFieldTouched(e.target.id, true, false);
   };
 
   const handleCodeVerification = (verificationCode?: string) => {
     const digitalAddressParams: SaveDigitalAddressParams = {
       addressType: AddressType.COURTESY,
       senderId: 'default',
-      channelType: type,
-      value: getContactValue(type),
+      channelType: ChannelType.EMAIL,
+      value: formik.values.email,
       code: verificationCode,
     };
 
@@ -113,7 +103,7 @@ const CourtesyContactItem = ({ type, value, blockDelete }: Props) => {
         dispatch(
           appStateActions.addSuccess({
             title: '',
-            message: t(`courtesy-contacts.${contactType}-added-successfully`, {
+            message: t(`courtesy-contacts.email-added-successfully`, {
               ns: 'recapiti',
             }),
           })
@@ -131,42 +121,49 @@ const CourtesyContactItem = ({ type, value, blockDelete }: Props) => {
     if (value) {
       digitalElemRef.current.toggleEdit();
     }
-    await formik.setFieldTouched(contactType, false, false);
-    await formik.setFieldValue(contactType, initialValues[contactType], true);
-  };
-
-  const handleChangeTouched = async (
-    event: ChangeEvent<HTMLTextAreaElement | HTMLInputElement>
-  ) => {
-    formik.handleChange(event);
-    await handleTouched(event.target.id, true);
-  };
-
-  const handleTouched = async (id: string, touched: boolean) => {
-    await formik.setFieldTouched(id, touched, false);
-  };
-
-  const handleEditConfirm = async (status: 'validated' | 'cancelled') => {
-    if (status === 'cancelled') {
-      await formik.setFieldValue(contactType, value, true);
-    }
+    await formik.setFieldTouched('email', false, false);
+    await formik.setFieldValue('email', initialValues.email, true);
   };
 
   const deleteConfirmHandler = () => {
     setModalOpen(null);
-    dispatch(
+    void dispatch(
       deleteAddress({
         addressType: AddressType.COURTESY,
         senderId: 'default',
-        channelType: type,
+        channelType: ChannelType.EMAIL,
       })
-    )
-      .unwrap()
-      .then(() => {
-        void handleTouched(contactType, false);
-      })
-      .catch(() => {});
+    );
   };
+
+  const handleAddressUpdateError = useCallback(
+    (responseError: AppResponse) => {
+      if (modalOpen === null) {
+        // notify the publisher we are not handling the error
+        return true;
+      }
+      if (Array.isArray(responseError.errors)) {
+        const error = responseError.errors[0];
+        codeModalRef.current?.updateError(
+          {
+            title: error.message.title,
+            content: error.message.content,
+          },
+          true
+        );
+      }
+      return false;
+    },
+    [modalOpen]
+  );
+
+  useEffect(() => {
+    AppResponsePublisher.error.subscribe('createOrUpdateAddress', handleAddressUpdateError);
+
+    return () => {
+      AppResponsePublisher.error.unsubscribe('createOrUpdateAddress', handleAddressUpdateError);
+    };
+  }, [handleAddressUpdateError]);
 
   /*
    * if *some* value (phone number, email address) has been attached to the contact type,
@@ -178,80 +175,63 @@ const CourtesyContactItem = ({ type, value, blockDelete }: Props) => {
    */
   return (
     <>
-      <form onSubmit={formik.handleSubmit} data-testid={`courtesyContacts-${contactType}`}>
-        <Typography id={`${contactType}-label`} variant="body2" mb={1} sx={{ fontWeight: 'bold' }}>
-          {t(`courtesy-contacts.${contactType}-added`, { ns: 'recapiti' })}
+      <form onSubmit={formik.handleSubmit} data-testid="courtesyContacts-email">
+        <Typography id="email-label" variant="body2" mb={1} sx={{ fontWeight: 'bold' }}>
+          {t(`courtesy-contacts.email-added`, { ns: 'recapiti' })}
         </Typography>
         {value ? (
           <DigitalContactElem
             senderId="default"
-            contactType={type}
+            contactType={ChannelType.EMAIL}
             ref={digitalElemRef}
             inputProps={{
-              id: contactType,
-              name: contactType,
-              label: t(`courtesy-contacts.link-${contactType}-placeholder`, {
+              id: 'email',
+              name: 'email',
+              label: t(`courtesy-contacts.link-email-placeholder`, {
                 ns: 'recapiti',
               }),
-              value: formik.values[contactType],
+              value: formik.values.email,
               onChange: (e) => void handleChangeTouched(e),
-              error:
-                (formik.touched[contactType] || formik.values[contactType].length > 0) &&
-                Boolean(formik.errors[contactType]),
-              helperText:
-                (formik.touched[contactType] || formik.values[contactType].length > 0) &&
-                formik.errors[contactType],
+              error: formik.touched.email && Boolean(formik.errors.email),
+              helperText: formik.touched.email && formik.errors.email,
             }}
             saveDisabled={!formik.isValid}
-            onConfirm={handleEditConfirm}
-            onEditCancel={() => handleEditConfirm('cancelled')}
             onDelete={() => setModalOpen(ModalType.DELETE)}
+            onEditCancel={() => formik.resetForm({ values: initialValues })}
             editManagedFromOutside
           />
         ) : (
           <Stack direction={{ xs: 'column', sm: 'row' }} spacing={2}>
             <TextField
-              id={contactType}
-              name={contactType}
-              aria-labelledby={`${contactType}-label`}
-              value={formik.values[contactType]}
+              id="email"
+              name="email"
+              value={formik.values.email}
               onChange={handleChangeTouched}
-              error={formik.touched[contactType] && Boolean(formik.errors[contactType])}
-              helperText={formik.touched[contactType] && formik.errors[contactType]}
+              error={formik.touched.email && Boolean(formik.errors.email)}
+              helperText={formik.touched.email && formik.errors.email}
               inputProps={{ sx: { height: '14px' } }}
-              placeholder={t(`courtesy-contacts.link-${contactType}-placeholder`, {
-                ns: 'recapiti',
-              })}
+              placeholder={t(`courtesy-contacts.link-email-placeholder`, { ns: 'recapiti' })}
               fullWidth
-              InputProps={
-                type === ChannelType.SMS
-                  ? {
-                      startAdornment: (
-                        <InputAdornment position="start">{internationalPhonePrefix}</InputAdornment>
-                      ),
-                    }
-                  : {}
-              }
               sx={{ flexBasis: { xs: 'unset', lg: '66.66%' } }}
             />
 
             <Button
-              id={`courtesy-${contactType}-button`}
+              id="courtesy-email-button"
               variant="outlined"
               disabled={!formik.isValid}
               fullWidth
               type="submit"
-              data-testid={`courtesy-${contactType}-button`}
+              data-testid="courtesy-email-button"
               sx={{ flexBasis: { xs: 'unset', lg: '33.33%' } }}
             >
-              {t(`courtesy-contacts.${contactType}-add`, { ns: 'recapiti' })}
+              {t(`courtesy-contacts.email-add`, { ns: 'recapiti' })}
             </Button>
           </Stack>
         )}
       </form>
       <ExistingContactDialog
         open={modalOpen === ModalType.EXISTING}
-        value={formik.values[contactType]}
+        value={formik.values.email}
         handleDiscard={() => setModalOpen(null)}
         handleConfirm={() => handleCodeVerification()}
       />
@@ -264,21 +244,18 @@ const CourtesyContactItem = ({ type, value, blockDelete }: Props) => {
         onCancel={() => setModalOpen(null)}
         confirmLabel={t('button.conferma')}
         checkboxLabel={t('button.capito')}
-        content={t(`alert-dialog-${type}`, { ns: 'recapiti' })}
+        content={t(`alert-dialog-email`, { ns: 'recapiti' })}
       />
       <CodeModal
-        title={
-          t(`courtesy-contacts.${contactType}-verify`, { ns: 'recapiti' }) +
-          ` ${formik.values[contactType]}`
-        }
-        subtitle={<Trans i18nKey={`courtesy-contacts.${contactType}-verify-descr`} ns="recapiti" />}
+        title={t(`courtesy-contacts.email-verify`, { ns: 'recapiti' }) + ` ${formik.values.email}`}
+        subtitle={<Trans i18nKey="courtesy-contacts.email-verify-descr" ns="recapiti" />}
         open={modalOpen === ModalType.CODE}
         initialValues={new Array(5).fill('')}
         codeSectionTitle={t(`courtesy-contacts.insert-code`, { ns: 'recapiti' })}
         codeSectionAdditional={
           <>
             <Typography variant="body2" display="inline">
-              {t(`courtesy-contacts.${contactType}-new-code`, { ns: 'recapiti' })}
+              {t(`courtesy-contacts.email-new-code`, { ns: 'recapiti' })}
               &nbsp;
             </Typography>
             <ButtonNaked
@@ -305,17 +282,13 @@ const CourtesyContactItem = ({ type, value, blockDelete }: Props) => {
       />
       <DeleteDialog
         showModal={modalOpen === ModalType.DELETE}
-        removeModalTitle={t(
-          `courtesy-contacts.${blockDelete ? 'block-' : ''}remove-${contactType}-title`,
-          { ns: 'recapiti' }
-        )}
-        removeModalBody={t(
-          `courtesy-contacts.${blockDelete ? 'block-' : ''}remove-${contactType}-message`,
-          {
-            value: formik.values[type],
-            ns: 'recapiti',
-          }
-        )}
+        removeModalTitle={t(`courtesy-contacts.${blockDelete ? 'block-' : ''}remove-email-title`, {
+          ns: 'recapiti',
+        })}
+        removeModalBody={t(`courtesy-contacts.${blockDelete ? 'block-' : ''}remove-email-message`, {
+          value: formik.values.email,
+          ns: 'recapiti',
+        })}
         handleModalClose={() => setModalOpen(null)}
         confirmHandler={deleteConfirmHandler}
         blockDelete={blockDelete}
@@ -324,4 +297,4 @@ const CourtesyContactItem = ({ type, value, blockDelete }: Props) => {
   );
 };
 
-export default CourtesyContactItem;
+export default EmailContactItem;
