+++ packages/pn-personagiuridica-webapp/src/components/Contacts/SmsContactItem.tsx
@@ -3,7 +3,7 @@ import { ChangeEvent, useCallback, useEffect, useRef, useState } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import * as yup from 'yup';
 
-import { Box, Button, Stack, TextField, Typography } from '@mui/material';
+import { Box, Button, InputAdornment, Stack, TextField, Typography } from '@mui/material';
 import {
   AppResponse,
   AppResponsePublisher,
@@ -14,14 +14,16 @@ import {
 } from '@pagopa-pn/pn-commons';
 import { ButtonNaked } from '@pagopa/mui-italia';
 
-import { PFEventsType } from '../../models/PFEventsType';
 import { AddressType, ChannelType } from '../../models/contacts';
 import { createOrUpdateAddress, deleteAddress } from '../../redux/contact/actions';
 import { SaveDigitalAddressParams } from '../../redux/contact/types';
 import { useAppDispatch, useAppSelector } from '../../redux/hooks';
 import { RootState } from '../../redux/store';
-import PFEventStrategyFactory from '../../utility/MixpanelUtils/PFEventStrategyFactory';
-import { contactAlreadyExists, emailValidationSchema } from '../../utility/contacts.utility';
+import {
+  contactAlreadyExists,
+  internationalPhonePrefix,
+  phoneValidationSchema,
+} from '../../utility/contacts.utility';
 import DeleteDialog from './DeleteDialog';
 import DigitalContactElem from './DigitalContactElem';
 import ExistingContactDialog from './ExistingContactDialog';
@@ -38,7 +40,7 @@ enum ModalType {
   DELETE = 'delete',
 }
 
-const EmailContactItem = ({ value, blockDelete }: Props) => {
+const SmsContactItem = ({ value, blockDelete }: Props) => {
   const { t } = useTranslation(['common', 'recapiti']);
   const digitalAddresses =
     useAppSelector((state: RootState) => state.contactsState.digitalAddresses) ?? [];
@@ -51,25 +53,25 @@ const EmailContactItem = ({ value, blockDelete }: Props) => {
   const codeModalRef =
     useRef<{ updateError: (error: ErrorMessage, codeNotValid: boolean) => void }>(null);
 
+  // value contains the prefix
+  const contactValue = value.replace(internationalPhonePrefix, '');
+
   const validationSchema = yup.object().shape({
-    email: emailValidationSchema(t),
+    sms: phoneValidationSchema(t),
   });
 
   const initialValues = {
-    email: value ?? '',
+    sms: contactValue ?? '',
   };
 
   const formik = useFormik({
     initialValues,
     validationSchema,
-    validateOnMount: true,
     enableReinitialize: true,
+    validateOnMount: true,
     onSubmit: () => {
-      PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_EMAIL_START, 'default');
       // first check if contact already exists
-      if (
-        contactAlreadyExists(digitalAddresses, formik.values.email, 'default', ChannelType.EMAIL)
-      ) {
+      if (contactAlreadyExists(digitalAddresses, formik.values.sms, 'default', ChannelType.SMS)) {
         setModalOpen(ModalType.EXISTING);
         return;
       }
@@ -83,15 +85,11 @@ const EmailContactItem = ({ value, blockDelete }: Props) => {
   };
 
   const handleCodeVerification = (verificationCode?: string) => {
-    if (verificationCode) {
-      PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_EMAIL_UX_CONVERSION, 'default');
-    }
-
     const digitalAddressParams: SaveDigitalAddressParams = {
       addressType: AddressType.COURTESY,
       senderId: 'default',
-      channelType: ChannelType.EMAIL,
-      value: formik.values.email,
+      channelType: ChannelType.SMS,
+      value: internationalPhonePrefix + formik.values.sms,
       code: verificationCode,
     };
 
@@ -101,19 +99,16 @@ const EmailContactItem = ({ value, blockDelete }: Props) => {
         // contact to verify
         // open code modal
         if (!res) {
-          // aprire la code modal
           setModalOpen(ModalType.CODE);
           return;
         }
-
-        PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_EMAIL_UX_SUCCESS, 'default');
 
         // contact has already been verified
         // show success message
         dispatch(
           appStateActions.addSuccess({
             title: '',
-            message: t(`courtesy-contacts.email-added-successfully`, {
+            message: t(`courtesy-contacts.sms-added-successfully`, {
               ns: 'recapiti',
             }),
           })
@@ -131,24 +126,19 @@ const EmailContactItem = ({ value, blockDelete }: Props) => {
     if (value) {
       digitalElemRef.current.toggleEdit();
     }
-    await formik.setFieldTouched('email', false, false);
-    await formik.setFieldValue('email', initialValues.email, true);
+    await formik.setFieldTouched('sms', false, false);
+    await formik.setFieldValue('sms', initialValues.sms, true);
   };
 
   const deleteConfirmHandler = () => {
     setModalOpen(null);
-    dispatch(
+    void dispatch(
       deleteAddress({
         addressType: AddressType.COURTESY,
         senderId: 'default',
-        channelType: ChannelType.EMAIL,
+        channelType: ChannelType.SMS,
       })
-    )
-      .unwrap()
-      .then(() => {
-        PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_REMOVE_EMAIL_SUCCESS, 'default');
-      })
-      .catch(() => {});
+    );
   };
 
   const handleAddressUpdateError = useCallback(
@@ -166,7 +156,6 @@ const EmailContactItem = ({ value, blockDelete }: Props) => {
           },
           true
         );
-        PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_EMAIL_CODE_ERROR);
       }
       return false;
     },
@@ -191,25 +180,26 @@ const EmailContactItem = ({ value, blockDelete }: Props) => {
    */
   return (
     <>
-      <form onSubmit={formik.handleSubmit} data-testid="courtesyContacts-email">
-        <Typography id="email-label" variant="body2" mb={1} sx={{ fontWeight: 'bold' }}>
-          {t(`courtesy-contacts.email-added`, { ns: 'recapiti' })}
+      <form onSubmit={formik.handleSubmit} data-testid="courtesyContacts-sms">
+        <Typography id="sms-label" variant="body2" mb={1} sx={{ fontWeight: 'bold' }}>
+          {t(`courtesy-contacts.sms-added`, { ns: 'recapiti' })}
         </Typography>
         {value ? (
           <DigitalContactElem
             senderId="default"
-            contactType={ChannelType.EMAIL}
+            contactType={ChannelType.SMS}
             ref={digitalElemRef}
             inputProps={{
-              id: 'email',
-              name: 'email',
-              label: t(`courtesy-contacts.link-email-placeholder`, {
+              id: 'sms',
+              name: 'sms',
+              label: t(`courtesy-contacts.link-sms-placeholder`, {
                 ns: 'recapiti',
               }),
-              value: formik.values.email,
+              value: formik.values.sms,
               onChange: (e) => void handleChangeTouched(e),
-              error: formik.touched.email && Boolean(formik.errors.email),
-              helperText: formik.touched.email && formik.errors.email,
+              error: formik.touched.sms && Boolean(formik.errors.sms),
+              helperText: formik.touched.sms && formik.errors.sms,
+              prefix: internationalPhonePrefix,
             }}
             saveDisabled={!formik.isValid}
             onDelete={() => setModalOpen(ModalType.DELETE)}
@@ -219,35 +209,42 @@ const EmailContactItem = ({ value, blockDelete }: Props) => {
         ) : (
           <Stack direction={{ xs: 'column', sm: 'row' }} spacing={2}>
             <TextField
-              id="email"
-              name="email"
-              value={formik.values.email}
+              id="sms"
+              name="sms"
+              value={formik.values.sms}
               onChange={handleChangeTouched}
-              error={formik.touched.email && Boolean(formik.errors.email)}
-              helperText={formik.touched.email && formik.errors.email}
+              error={formik.touched.sms && Boolean(formik.errors.sms)}
+              helperText={formik.touched.sms && formik.errors.sms}
               inputProps={{ sx: { height: '14px' } }}
-              placeholder={t(`courtesy-contacts.link-email-placeholder`, { ns: 'recapiti' })}
+              placeholder={t(`courtesy-contacts.link-sms-placeholder`, {
+                ns: 'recapiti',
+              })}
               fullWidth
+              InputProps={{
+                startAdornment: (
+                  <InputAdornment position="start">{internationalPhonePrefix}</InputAdornment>
+                ),
+              }}
               sx={{ flexBasis: { xs: 'unset', lg: '66.66%' } }}
             />
 
             <Button
-              id="courtesy-email-button"
+              id="courtesy-sms-button"
               variant="outlined"
               disabled={!formik.isValid}
               fullWidth
               type="submit"
-              data-testid="courtesy-email-button"
+              data-testid="courtesy-sms-button"
               sx={{ flexBasis: { xs: 'unset', lg: '33.33%' } }}
             >
-              {t(`courtesy-contacts.email-add`, { ns: 'recapiti' })}
+              {t(`courtesy-contacts.sms-add`, { ns: 'recapiti' })}
             </Button>
           </Stack>
         )}
       </form>
       <ExistingContactDialog
         open={modalOpen === ModalType.EXISTING}
-        value={formik.values.email}
+        value={formik.values.sms}
         handleDiscard={() => setModalOpen(null)}
         handleConfirm={() => handleCodeVerification()}
       />
@@ -260,18 +257,18 @@ const EmailContactItem = ({ value, blockDelete }: Props) => {
         onCancel={() => setModalOpen(null)}
         confirmLabel={t('button.conferma')}
         checkboxLabel={t('button.capito')}
-        content={t(`alert-dialog-email`, { ns: 'recapiti' })}
+        content={t(`alert-dialog-sms`, { ns: 'recapiti' })}
       />
       <CodeModal
-        title={t(`courtesy-contacts.email-verify`, { ns: 'recapiti' }) + ` ${formik.values.email}`}
-        subtitle={<Trans i18nKey="courtesy-contacts.email-verify-descr" ns="recapiti" />}
+        title={t(`courtesy-contacts.sms-verify`, { ns: 'recapiti' }) + ` ${formik.values.sms}`}
+        subtitle={<Trans i18nKey="courtesy-contacts.sms-verify-descr" ns="recapiti" />}
         open={modalOpen === ModalType.CODE}
         initialValues={new Array(5).fill('')}
         codeSectionTitle={t(`courtesy-contacts.insert-code`, { ns: 'recapiti' })}
         codeSectionAdditional={
           <>
             <Typography variant="body2" display="inline">
-              {t(`courtesy-contacts.email-new-code`, { ns: 'recapiti' })}
+              {t(`courtesy-contacts.sms-new-code`, { ns: 'recapiti' })}
               &nbsp;
             </Typography>
             <ButtonNaked
@@ -298,11 +295,11 @@ const EmailContactItem = ({ value, blockDelete }: Props) => {
       />
       <DeleteDialog
         showModal={modalOpen === ModalType.DELETE}
-        removeModalTitle={t(`courtesy-contacts.${blockDelete ? 'block-' : ''}remove-email-title`, {
+        removeModalTitle={t(`courtesy-contacts.${blockDelete ? 'block-' : ''}remove-sms-title`, {
           ns: 'recapiti',
         })}
-        removeModalBody={t(`courtesy-contacts.${blockDelete ? 'block-' : ''}remove-email-message`, {
-          value: formik.values.email,
+        removeModalBody={t(`courtesy-contacts.${blockDelete ? 'block-' : ''}remove-sms-message`, {
+          value: formik.values.sms,
           ns: 'recapiti',
         })}
         handleModalClose={() => setModalOpen(null)}
@@ -313,4 +310,4 @@ const EmailContactItem = ({ value, blockDelete }: Props) => {
   );
 };
 
-export default EmailContactItem;
+export default SmsContactItem;
