+++ packages/pn-personagiuridica-webapp/src/components/Contacts/__test__/SmsContactItem.test.tsx
